+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 * SERVER: .KP TLD no longer provides public WHOIS interface (GH-52).
 
+* NEW: Added full whois.dreamhost.com parser.
+
 * FIXED: whois.nic.gl parser must support `Delegate' status.
 
 * FIXED: whois.godaddy.com partially changed response format (GH-148).